add another functor to be able to pass a map to ListPositionFunction
