+++ src/include/duckdb/function/scalar/nested_functions.hpp
@@ -14,11 +14,58 @@
 #include "duckdb/common/unordered_map.hpp"
 #include "duckdb/common/field_writer.hpp"
 #include "duckdb/function/built_in_functions.hpp"
+#include "duckdb/function/scalar/list/contains_or_position.hpp"
 
 namespace duckdb {
 
 enum class MapInvalidReason : uint8_t { VALID, NULL_KEY_LIST, NULL_KEY, DUPLICATE_KEY };
 enum class UnionInvalidReason : uint8_t { VALID, TAG_OUT_OF_RANGE, NO_MEMBERS, VALIDITY_OVERLAP };
+
+struct ListArgFunctor {
+	static Vector &GetList(Vector &list) {
+		return list;
+	}
+	static idx_t GetListSize(Vector &list) {
+		return ListVector::GetListSize(list);
+	}
+	static Vector &GetEntry(Vector &list) {
+		return ListVector::GetEntry(list);
+	}
+};
+
+struct MapKeyArgFunctor {
+	// MAP is a LIST(STRUCT(K,V))
+	// meaning the MAP itself is a List, but the child vector that we're interested in (the keys)
+	// are a level deeper than the initial child vector
+
+	static Vector &GetList(Vector &map) {
+		return map;
+	}
+	static idx_t GetListSize(Vector &map) {
+		return ListVector::GetListSize(map);
+	}
+	static Vector &GetEntry(Vector &map) {
+		return MapVector::GetKeys(map);
+	}
+};
+
+struct ContainsFunctor {
+	static inline bool Initialize() {
+		return false;
+	}
+	static inline bool UpdateResultEntries(idx_t child_idx) {
+		return true;
+	}
+};
+
+struct PositionFunctor {
+	static inline int32_t Initialize() {
+		return 0;
+	}
+	static inline int32_t UpdateResultEntries(idx_t child_idx) {
+		return child_idx + 1;
+	}
+};
 
 struct VariableReturnBindData : public FunctionData {
 	LogicalType stype;
@@ -125,7 +172,6 @@ struct ListFlattenFun {
 };
 
 struct ListPositionFun {
-	static void ListPositionFunction(DataChunk &args, ExpressionState &state, Vector &result);
 	static ScalarFunction GetFunction();
 	static void RegisterFunction(BuiltinFunctions &set);
 };
